SOV-108: PR review fixes
