+++ apps/frontend/package.json
@@ -6,9 +6,9 @@
     "@sovryn/ethers-provider": "*",
     "@sovryn/tailwindcss-config": "*",
     "@sovryn/ui": "*",
-    "ethers": "^5.7.1",
-    "react": "^18.1.0",
-    "react-dom": "^18.1.0",
+    "ethers": "5.7.1",
+    "react": "18.1.0",
+    "react-dom": "18.1.0",
     "react-scripts": "5.0.1"
   },
   "devDependencies": {
