+++ packages/ethers-provider/package.json
@@ -16,18 +16,18 @@
     "lint": "eslint ./src/**/*.ts"
   },
   "devDependencies": {
-    "@babel/preset-typescript": "^7.18.6",
-    "jest": "^29.0.3",
-    "tsup": "^6.2.3",
-    "typescript": "^4.8.3"
+    "@babel/preset-typescript": "7.18.6",
+    "jest": "29.0.3",
+    "tsup": "6.2.3",
+    "typescript": "4.8.3"
   },
   "files": [
     "dist"
   ],
   "dependencies": {
-    "rxjs": "^7.5.6"
+    "rxjs": "7.5.6"
   },
   "peerDependencies": {
-    "ethers": "^5.7"
+    "ethers": "5.7"
   }
 }
